HEEDLS-664 Renaming and refactoring to prevent merge conflicts
