+++ DigitalLearningSolutions.Data.Tests/Services/ActionPlanServiceTests.cs
@@ -265,6 +265,21 @@
         }
 
         [Test]
+        public void SetCompletionDate_calls_data_service()
+        {
+            // Given
+            const int learningLogItemId = 1;
+            var completedDate = new DateTime(2021, 09, 01);
+
+            // When
+            learningLogItemsDataService.SetCompletionDate(learningLogItemId, completedDate);
+
+            // Then
+            A.CallTo(() => learningLogItemsDataService.SetCompletionDate(learningLogItemId, completedDate))
+                .MustHaveHappenedOnceExactly();
+        }
+
+        [Test]
         public void RemoveActionPlanResource_removes_item()
         {
             // Given
