+++ DigitalLearningSolutions.Web/Views/LearningPortal/Current/_CurrentLearningResourceCard.cshtml
@@ -32,7 +32,7 @@
       </div>
       <a class="nhsuk-button nhsuk-button--secondary small-edit-button nhsuk-u-margin-bottom-4"
          aria-describedby="@Model.Id-name-lhr"
-         asp-action="MarkCurrentLearningLogItemAsComplete"
+         asp-action="MarkCurrentActionPlanItemAsComplete"
          asp-route-learningLogItemId="@Model.Id">
          role="button">
         Mark as complete
